Corrected the voucher locking functionality.

+++ gkcore/views/api_transaction.py
@@ -504,21 +504,22 @@ class api_transaction(object):
 				self.con = eng.connect()
 				dataset = self.request.json_body
 				result = self.con.execute(vouchers.update().where(vouchers.c.vouchercode==dataset["vouchercode"]).values(dataset))
-				drs = dataset["drs"]
-				crs = dataset["crs"]
-				delrecoresult = self.con.execute("delete from bankrecon where vouchercode = %d"%(int(dataset["vouchercode"])))
-				for drkeys in drs.keys():
-					accgrpdata = self.con.execute(select([groupsubgroups.c.groupname,groupsubgroups.c.groupcode]).where(groupsubgroups.c.groupcode==(select([accounts.c.groupcode]).where(accounts.c.accountcode==int(drkeys)))))
-					accgrp = accgrpdata.fetchone()
-					if accgrp["groupname"] == "Bank":
-						vouchercode =dataset["vouchercode"]
-						recoresult = self.con.execute(bankrecon.insert(),[{"vouchercode":int(vouchercode),"accountcode":drkeys,"orgcode":authDetails["orgcode"]}])
-				for crkeys in crs.keys():
-					accgrpdata = self.con.execute(select([groupsubgroups.c.groupname,groupsubgroups.c.groupcode]).where(groupsubgroups.c.groupcode==(select([accounts.c.groupcode]).where(accounts.c.accountcode==int(crkeys)))))
-					accgrp = accgrpdata.fetchone()
-					if accgrp["groupname"] == "Bank":
-						vouchercode =dataset["vouchercode"]
-						recoresult = self.con.execute(bankrecon.insert(),[{"vouchercode":int(vouchercode),"accountcode":crkeys,"orgcode":authDetails["orgcode"]}])
+				if dataset.has_key("drs"):
+					drs = dataset["drs"]
+					crs = dataset["crs"]
+					delrecoresult = self.con.execute("delete from bankrecon where vouchercode = %d"%(int(dataset["vouchercode"])))
+					for drkeys in drs.keys():
+						accgrpdata = self.con.execute(select([groupsubgroups.c.groupname,groupsubgroups.c.groupcode]).where(groupsubgroups.c.groupcode==(select([accounts.c.groupcode]).where(accounts.c.accountcode==int(drkeys)))))
+						accgrp = accgrpdata.fetchone()
+						if accgrp["groupname"] == "Bank":
+							vouchercode =dataset["vouchercode"]
+							recoresult = self.con.execute(bankrecon.insert(),[{"vouchercode":int(vouchercode),"accountcode":drkeys,"orgcode":authDetails["orgcode"]}])
+					for crkeys in crs.keys():
+						accgrpdata = self.con.execute(select([groupsubgroups.c.groupname,groupsubgroups.c.groupcode]).where(groupsubgroups.c.groupcode==(select([accounts.c.groupcode]).where(accounts.c.accountcode==int(crkeys)))))
+						accgrp = accgrpdata.fetchone()
+						if accgrp["groupname"] == "Bank":
+							vouchercode =dataset["vouchercode"]
+							recoresult = self.con.execute(bankrecon.insert(),[{"vouchercode":int(vouchercode),"accountcode":crkeys,"orgcode":authDetails["orgcode"]}])
 				self.con.close()
 				return {"gkstatus":enumdict["Success"]}
 			#except:
